test: Move buildbot.{test.unit => test}.runprocess
