+++ master/buildbot/test/unit/test_machine_generic.py
@@ -27,9 +27,9 @@ from buildbot.machine.generic import RemoteSshWakeAction
 from buildbot.machine.generic import RemoteSshWOLAction
 from buildbot.test.fake.private_tempdir import MockPrivateTemporaryDirectory
 from buildbot.test.reactor import TestReactorMixin
+from buildbot.test.runprocess import ExpectMasterShell
+from buildbot.test.runprocess import MasterRunProcessMixin
 from buildbot.test.util import config
-from buildbot.test.util.runprocess import ExpectMasterShell
-from buildbot.test.util.runprocess import MasterRunProcessMixin
 
 
 class FakeManager:
