+++ master/buildbot/test/unit/test_secret_in_passwordstore.py
@@ -23,9 +23,9 @@ from twisted.trial import unittest
 from buildbot.secrets.providers.passwordstore import SecretInPass
 from buildbot.test.fake import fakemaster
 from buildbot.test.reactor import TestReactorMixin
+from buildbot.test.runprocess import ExpectMasterShell
+from buildbot.test.runprocess import MasterRunProcessMixin
 from buildbot.test.util.config import ConfigErrorsMixin
-from buildbot.test.util.runprocess import ExpectMasterShell
-from buildbot.test.util.runprocess import MasterRunProcessMixin
 
 
 class TestSecretInPass(MasterRunProcessMixin, TestReactorMixin, ConfigErrorsMixin,
