+++ master/buildbot/test/unit/worker/test_libvirt.py
@@ -26,8 +26,8 @@ from buildbot import config
 from buildbot.interfaces import LatentWorkerFailedToSubstantiate
 from buildbot.test.fake import libvirt as libvirtfake
 from buildbot.test.reactor import TestReactorMixin
-from buildbot.test.util.runprocess import ExpectMasterShell
-from buildbot.test.util.runprocess import MasterRunProcessMixin
+from buildbot.test.runprocess import ExpectMasterShell
+from buildbot.test.runprocess import MasterRunProcessMixin
 from buildbot.test.util.warnings import assertProducesWarnings
 from buildbot.warnings import DeprecatedApiWarning
 from buildbot.worker import libvirt as libvirtworker
